Updating code to use Gateway.
